fix: change dtype for `open_image` to `np.uint8`
The `rgb2lab` function from skimage is detecting whether the RGB input is between 0-1 or 0-255 by determining whether or not it is a float or an integer, and not also checking the range of values too. That means it’s detecting that the image is a float and it’s using that to calculate the LAB values, meaning the values are just wild. 

I changed the input `dtype` to `np.uint8` i.e. to an 8-bit integer (which is what RGB is) and it works now.

+++ ayesaac/services/object_detection/ColorDetection.py
@@ -19,7 +19,7 @@ class ColorDetection:
     @staticmethod
     def open_image(file_path: str):
         image = Image.open(file_path)
-        return np.array(image, dtype=np.float32)
+        return np.array(image, dtype=np.uint8)
     
     @staticmethod
     def convert_rgb_to_lab(image: np.ndarray) -> np.ndarray:
